+++ mini-spring/src/main/java/instantiationstrategy/beans/factory/support/AbstractAutowireCapableBeanFactory.java
@@ -1,0 +1,41 @@
+package instantiationstrategy.beans.factory.support;
+
+import instantiationstrategy.beans.BeansException;
+import instantiationstrategy.beans.factory.config.BeanDefinition;
+
+/**
+ * @Author huabin
+ * @DateTime 2022-11-30 09:57
+ * @Desc
+ */
+public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory {
+
+    private InstantiationStrategy instantiationStrategy = new SimpleInstantiationStrategy();
+
+    @Override
+    protected Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException {
+        return doCreateBean(beanName, beanDefinition);
+    }
+
+    protected Object doCreateBean(String beanName, BeanDefinition beanDefinition) {
+
+        Object bean = null;
+        try {
+            bean = createBeanInstance(beanDefinition);
+        } catch (Exception e) {
+            throw new BeansException("Instantiation of bean failed", e);
+        }
+
+        addSingleton(beanName, bean);
+        return bean;
+    }
+
+    private Object createBeanInstance(BeanDefinition beanDefinition) {
+
+        return getInstantiationStrategy().instantiate(beanDefinition);
+    }
+
+    public InstantiationStrategy getInstantiationStrategy(){
+        return instantiationStrategy;
+    }
+}