+++ mini-spring/src/main/java/instantiationstrategy/beans/factory/support/DefaultListableBeanFactory.java
@@ -1,0 +1,32 @@
+package instantiationstrategy.beans.factory.support;
+
+import instantiationstrategy.beans.BeansException;
+import instantiationstrategy.beans.factory.config.BeanDefinition;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * @Author huabin
+ * @DateTime 2022-11-30 09:42
+ * @Desc
+ */
+public class DefaultListableBeanFactory extends AbstractAutowireCapableBeanFactory implements BeanDefinitionRegistry {
+
+    private Map<String, BeanDefinition> beanDefinitionMap = new HashMap<>();
+
+    @Override
+    public void registerBeanDefinition(String beanName, BeanDefinition beanDefinition) {
+        beanDefinitionMap.put(beanName, beanDefinition);
+    }
+
+    @Override
+    protected BeanDefinition getBeanDefinition(String beanName) throws Exception {
+        BeanDefinition beanDefinition = beanDefinitionMap.get(beanName);
+
+        if (beanDefinition == null) {
+            throw new BeansException("No bean named '" + beanName + "' is defined");
+        }
+        return beanDefinition;
+    }
+}