+++ mini-spring/src/main/java/instantiationstrategy/beans/factory/support/InstantiationStrategy.java
@@ -1,0 +1,10 @@
+package instantiationstrategy.beans.factory.support;
+
+import instantiationstrategy.beans.BeansException;
+import instantiationstrategy.beans.factory.config.BeanDefinition;
+
+public interface InstantiationStrategy {
+
+    Object instantiate(BeanDefinition beanDefinition) throws BeansException;
+
+}