+++ mini-spring/src/test/java/instantiationstrategy/HelloService.java
@@ -1,0 +1,15 @@
+package instantiationstrategy;
+
+/**
+ * @Author huabin
+ * @DateTime 2022-11-30 10:21
+ * @Desc
+ */
+public class HelloService {
+
+    public String sayHello() {
+        System.out.println("hello");
+        return "hello";
+    }
+
+}