+++ mini-spring/changelog.md
@@ -12,3 +12,11 @@
 - BeanDefinitionRegistry，BeanDefinition注册表接口，定义注册BeanDefinition的方法。
 - SingletonBeanRegistry及其实现类DefaultSingletonBeanRegistry，定义添加和获取单例bean的方法。
 bean容器作为BeanDefinitionRegistry和SingletonBeanRegistry的实现类，具备两者的能力。向bean容器中注册BeanDefinition后，使用bean时才会实例化。
+
+## Bean实例化策略InstantiationStrategy
+> package:instantiationstrategy
+
+之前创建bean的方法是在AbstractAutowireCapableBeanFactory#doCreateBean中用newInstance方法实现的，仅适用于bean有无参构造的情况。
+针对bean的实例化，抽象出一个实例化策略的接口InstantiationStrategy，有两个实现类：
+- SimpleInstantiationStrategy，使用bean的构造函数来实例化
+- CglibSubclassingInstantiationStrategy，使用CGLIB动态生成子类